fix: logs not being rendered when clearing filters
