+++ app/src/components/mode-specific/desktop/InterceptTraffic/WebTraffic/TrafficTableV2/NetworkInspector/NetworkTable/VirtualTableV2.tsx
@@ -8,6 +8,7 @@ import { useVirtualizer } from "@tanstack/react-virtual";
 import "./virtualTableV2.css";
 import { Button } from "antd";
 import { ArrowDownOutlined } from "@ant-design/icons";
+import NoTrafficCTA from "./NoTrafficCTA";
 
 interface Props {
   logs: any;
@@ -17,15 +18,16 @@ interface Props {
   onReplayRequest: () => void;
 }
 
-const VirtualTableV2 = ({ logs, renderHeader, renderLogRow, selectedRowData, onReplayRequest }: Props) => {
+const VirtualTableV2 = ({ logs = [], renderHeader, renderLogRow, selectedRowData, onReplayRequest }: Props) => {
   const [selected, setSelected] = useState(null);
   const [lastKnowBottomIndex, setLastKnownBottomIndex] = useState(null);
   const [isScrollToBottomEnabled, setIsScrollToBottomEnabled] = useState(true);
+  const [logsLength, setLogsLength] = useState(logs.length);
 
   const parentRef = useRef(null);
 
   const rowVirtualizer = useVirtualizer({
-    count: logs.length,
+    count: logsLength,
     getScrollElement: () => parentRef.current,
     estimateSize: () => ITEM_SIZE,
     onChange: (virtualizer) => {
@@ -63,9 +65,9 @@ const VirtualTableV2 = ({ logs, renderHeader, renderLogRow, selectedRowData, onR
   };
 
   const scrollToBottom = useCallback(() => {
-    rowVirtualizer.scrollToIndex(logs.length - 1, { align: "start" });
-    // rowVirtualizer.scrollToOffset(rowVirtualizer.getTotalSize(), {align: 'start'});
-  }, [logs, rowVirtualizer]);
+    if (logsLength > 0) rowVirtualizer.scrollToIndex(logsLength - 1, { align: "start" });
+    // rowVirtualizer.scrollToOffset(rowVirtualizer.getTotalSize(), { align: "start" });
+  }, [logsLength, rowVirtualizer]);
 
   useEffect(() => {
     if (isScrollToBottomEnabled) {
@@ -73,7 +75,7 @@ const VirtualTableV2 = ({ logs, renderHeader, renderLogRow, selectedRowData, onR
     }
   }, [scrollToBottom, isScrollToBottomEnabled]);
 
-  const renderNewLogsButton = () => {
+  const renderNewLogsButton = useCallback(() => {
     if (isScrollToBottomEnabled) {
       return null;
     }
@@ -82,7 +84,7 @@ const VirtualTableV2 = ({ logs, renderHeader, renderLogRow, selectedRowData, onR
 
     let newLogsCount = 0;
     if (lastKnowBottomIndex) {
-      newLogsCount = logs.length - 1 - lastKnowBottomIndex;
+      newLogsCount = (logsLength > 0 ? logsLength - 1 : logsLength) - lastKnowBottomIndex;
       buttonText = newLogsCount > 1 ? `${newLogsCount} new logs` : "1 new log";
     }
 
@@ -100,7 +102,15 @@ const VirtualTableV2 = ({ logs, renderHeader, renderLogRow, selectedRowData, onR
     }
 
     return null;
-  };
+  }, [isScrollToBottomEnabled, lastKnowBottomIndex, logsLength, scrollToBottom]);
+
+  useEffect(() => {
+    setLogsLength(logs.length);
+  }, [logs, setLogsLength]);
+
+  if (logs.length === 0) {
+    return <NoTrafficCTA />;
+  }
 
   return (
     <>
